+++ src/main/java/com/donikrizky/kicau/authservice/service/AuthService.java
@@ -5,7 +5,7 @@ import com.donikrizky.kicau.authservice.dto.request.RegisterUserRequestDTO;
 public interface AuthService {
 	public boolean login(String email, String password);
 
-	public boolean logout(Long userId);
+	public boolean logout(Integer userId);
 
 	public void save(RegisterUserRequestDTO requestDTO);
 }
